fix: change all Instant fields underlying database types
Having an java.time.Instant field mapped to a "TIMESTAMP WITHOUT TIME ZONE"
field was seeding the database improperly. Essentially, the database was seeing
a time stamp such as "2017-09-25T18:00:00Z" and stripping the "Z", since it is
"WITHOUT TIME ZONE", converting it to a local time stamp and storing that.